Isolate examples from storeis
