+++ packages/folo-values/src/stories/FormsSubmit.stories.js
@@ -1,50 +1,14 @@
 /* eslint-disable jsx-a11y/label-has-associated-control */
 import React from "react";
 
-import Field from "../components/Field";
 import Form from "../components/Form";
 
-const styleForm = {
-  display: "flex",
-  flexDirection: "column",
-  background: "beige",
-  width: "40%",
-  margin: "10px",
-  padding: "17px",
-};
-
-const styleLabel = {
-  padding: "17px",
-};
-
-const BasicForm = ({ onSubmit }) => (
-  <Form style={styleForm} onSubmit={onSubmit}>
-    type a text:
-    <Field valueKey="textInput" type="text" placeholder="start typing..." />
-    choose this checkbox:
-    <Field
-      valueKey="checkbox1"
-      type="checkbox"
-      groupName="test"
-      initValue={false}
-    />
-    or this checkbox:
-    <Field valueKey="checkbox2" type="checkbox" groupName="test" />
-    <label style={styleLabel}>
-      items
-      <Field valueKey="options" type="select">
-        <option>A</option>
-        <option>B</option>
-        <option>C</option>
-      </Field>
-    </label>
-    <button type="submit">submit</button>
-  </Form>
-);
+import BasicForm from "./examples/BasicForm";
+import AddressForm from "./examples/AddressForm";
 
 export default {
-  title: "Forms/Basic Submit",
-  component: BasicForm,
+  title: "Forms/Forms with Submit",
+  component: Form,
   argTypes: {
     onSubmit: {
       action: "onSubmit",
@@ -52,6 +16,6 @@ export default {
   },
 };
 
-const Template = (args) => <BasicForm {...args} />;
+export const SimpleForm = (args) => <BasicForm {...args} />;
 
-export const Basic = Template.bind({});
+export const CustomComponents = (args) => <AddressForm {...args} />;
