+++ packages/folo-values/src/stories/examples/AddressForm.js
@@ -1,6 +1,8 @@
 /* eslint-disable react/jsx-wrap-multilines */
 /* eslint-disable import/no-extraneous-dependencies */
+
 import React from "react";
+
 import Grid from "@material-ui/core/Grid";
 import Typography from "@material-ui/core/Typography";
 import TextField from "@material-ui/core/TextField";
@@ -8,10 +10,10 @@ import FormControlLabel from "@material-ui/core/FormControlLabel";
 import Checkbox from "@material-ui/core/Checkbox";
 import Button from "@material-ui/core/Button";
 
-import Field from "../components/Field";
-import Form from "../components/Form";
+import Field from "../../components/Field";
+import Form from "../../components/Form";
 
-export function AddressForm({ onSubmit }) {
+const AddressForm = ({ onSubmit }) => {
   return (
     <div
       style={{
@@ -147,14 +149,6 @@ export function AddressForm({ onSubmit }) {
       </div>
     </div>
   );
-}
-
-export default {
-  title: "Forms/Custom components",
-  component: AddressForm,
-  argTypes: {
-    onSubmit: {
-      action: "onSubmit",
-    },
-  },
 };
+
+export default AddressForm;
